Simplification and switching to PyCharm
PyCharm is loved by many. Looks very good. 1. No more need for symbolic
links in the scripts folder for now. PyCharm works out of the gate
there. Eventually they should be integrated as actual tests in the
package. Maybe have more of a problem with iPython. But trying to forget
about that right now and work just in scripts. 2. Killing of test
iPython notebooks. I know it works, but for right now I am getting
distracted.